// react/components/api/checkoutexternal/models/CustomPlanActivationStrategy.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Schematic API
 * Schematic API
 *
 * The version of the OpenAPI document: 0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const CustomPlanActivationStrategy = {
    Payment: 'on_payment',
    Publish: 'on_publish'
} as const;
export type CustomPlanActivationStrategy = typeof CustomPlanActivationStrategy[keyof typeof CustomPlanActivationStrategy];


export function instanceOfCustomPlanActivationStrategy(value: any): boolean {
    for (const key in CustomPlanActivationStrategy) {
        if (Object.prototype.hasOwnProperty.call(CustomPlanActivationStrategy, key)) {
            if (CustomPlanActivationStrategy[key as keyof typeof CustomPlanActivationStrategy] === value) {
                return true;
            }
        }
    }
    return false;
}

export function CustomPlanActivationStrategyFromJSON(json: any): CustomPlanActivationStrategy {
    return CustomPlanActivationStrategyFromJSONTyped(json, false);
}

export function CustomPlanActivationStrategyFromJSONTyped(json: any, ignoreDiscriminator: boolean): CustomPlanActivationStrategy {
    return json as CustomPlanActivationStrategy;
}

export function CustomPlanActivationStrategyToJSON(value?: CustomPlanActivationStrategy | null): any {
    return value as any;
}

export function CustomPlanActivationStrategyToJSONTyped(value: any, ignoreDiscriminator: boolean): CustomPlanActivationStrategy {
    return value as CustomPlanActivationStrategy;
}