// react/components/api/checkoutexternal/models/BillingSubscriptionView.ts
export interface BillingSubscriptionView {
    /**
     * 
     * @type {string}
     * @memberof BillingSubscriptionView
     */
    applicationId?: string | null;
    /**
     * 
     * @type {number}
     * @memberof BillingSubscriptionView
     */
    cancelAt?: number | null;
    /**
     * 
     * @type {boolean}
     * @memberof BillingSubscriptionView
     */
    cancelAtPeriodEnd: boolean;
    /**
     * 
     * @type {string}
     * @memberof BillingSubscriptionView
     */
    companyId?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof BillingSubscriptionView
     */
    createdAt: Date;
    /**
     * 
     * @type {string}
     * @memberof BillingSubscriptionView
     */
    currency: string;
    /**
     * 
     * @type {string}
     * @memberof BillingSubscriptionView
     */
    customerExternalId: string;
    /**
     * 
     * @type {string}
     * @memberof BillingSubscriptionView
     */
    defaultPaymentMethodId?: string | null;
    /**
     * 
     * @type {Array<BillingSubscriptionDiscountView>}
     * @memberof BillingSubscriptionView
     */
    discounts: Array<BillingSubscriptionDiscountView>;
    /**
     * 
     * @type {Date}
     * @memberof BillingSubscriptionView
     */
    expiredAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof BillingSubscriptionView
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof BillingSubscriptionView
     */
    interval: string;
    /**
     * 
     * @type {InvoiceResponseData}
     * @memberof BillingSubscriptionView
     */
    latestInvoice?: InvoiceResponseData;
    /**
     * 
     * @type {object}
     * @memberof BillingSubscriptionView
     */
    metadata?: object;
    /**
     * 
     * @type {PaymentMethodResponseData}
     * @memberof BillingSubscriptionView
     */
    paymentMethod?: PaymentMethodResponseData;
    /**
     * 
     * @type {number}
     * @memberof BillingSubscriptionView
     */
    periodEnd: number;
    /**
     * 
     * @type {number}
     * @memberof BillingSubscriptionView
     */
    periodStart: number;
    /**
     * 
     * @type {Array<BillingProductForSubscriptionResponseData>}
     * @memberof BillingSubscriptionView
     */
    products: Array<BillingProductForSubscriptionResponseData>;
    /**
     * 
     * @type {BillingProviderType}
     * @memberof BillingSubscriptionView
     */
    providerType: BillingProviderType;
    /**
     * 
     * @type {string}
     * @memberof BillingSubscriptionView
     */
    status: string;
    /**
     * 
     * @type {string}
     * @memberof BillingSubscriptionView
     */
    subscriptionExternalId: string;
    /**
     * 
     * @type {number}
     * @memberof BillingSubscriptionView
     */
    totalPrice: number;
    /**
     * 
     * @type {number}
     * @memberof BillingSubscriptionView
     */
    trialEnd?: number | null;
    /**
     * 
     * @type {BillingSubscriptionTrialEndSetting}
     * @memberof BillingSubscriptionView
     */
    trialEndSetting?: BillingSubscriptionTrialEndSetting | null;
}