// react/components/api/checkoutexternal/models/CustomPlanBillingResponseData.ts
export interface CustomPlanBillingResponseData {
    /**
     * 
     * @type {CustomPlanActivationStrategy}
     * @memberof CustomPlanBillingResponseData
     */
    activationStrategy: CustomPlanActivationStrategy;
    /**
     * 
     * @type {string}
     * @memberof CustomPlanBillingResponseData
     */
    companyId: string;
    /**
     * 
     * @type {Date}
     * @memberof CustomPlanBillingResponseData
     */
    createdAt: Date;
    /**
     * 
     * @type {number}
     * @memberof CustomPlanBillingResponseData
     */
    daysUntilDue: number;
    /**
     * 
     * @type {string}
     * @memberof CustomPlanBillingResponseData
     */
    externalInvoiceId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof CustomPlanBillingResponseData
     */
    id: string;
    /**
     * 
     * @type {Date}
     * @memberof CustomPlanBillingResponseData
     */
    paidAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof CustomPlanBillingResponseData
     */
    planId: string;
    /**
     * 
     * @type {Date}
     * @memberof CustomPlanBillingResponseData
     */
    publishedAt?: Date | null;
    /**
     * 
     * @type {CustomPlanBillingStatus}
     * @memberof CustomPlanBillingResponseData
     */
    status: CustomPlanBillingStatus;
    /**
     * 
     * @type {string}
     * @memberof CustomPlanBillingResponseData
     */
    stripeInvoiceUrl?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof CustomPlanBillingResponseData
     */
    updatedAt: Date;
}